Fix whitespace lexing/parsing.
Our consumeWhitespace method had a bug where it wouldn't emit the whitespace
token, meaning we weren't actually skipping the whitespace. Oops.

While in there, I removed an outdated TODO (we already supplied a default
state).

I updated the lexing of unquoted strings to take into account the rules about
spaces and unquoted strings; basically, spaces before or after are ignored as
whitespace, spaces in the middle ocunt, and empty strings are not allowed.

I removed an extra case when detecting what to do when lexing an unquoted
string; we already consumed all the whitespace, so the next character shouldn't
be whitespace, so no need to test for it.

We need to consume whitespace before we start lexing the separator character.

I updated the token debugging to be a bit more useful, by defining a String()
method on the token type itself, so it'll expose both the type and the value.
This makes unexpected errors easier to deal with, and is used in all the errors
raised by the parser now.

I added a bunch of whitespace tests for lexing and parsing.